pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "ifpdf"
version = "0.1.0"
description = "Convert PDFs into AI-friendly Markdown"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "You"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pymupdf>=1.23",
    "pdfplumber>=0.10",
    "tiktoken>=0.5",
    "pyperclip>=1.8",
    "pillow>=10.0",
]

[project.optional-dependencies]
ocr = [
    "pdf2image>=1.16",
    "pytesseract>=0.3",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.scripts]
ifpdf = "ifpdf.cli:app"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.pytest.ini_options]
testpaths = ["tests"]
